requirements.txt

argparse
cupy
opencv-python-headless
glob
gryds
imageio
matplotlib
numpy
pandas
scipy
SimpleITK
scikit-learn
torch
torchvision
tqdm
pyyaml
